Saathi/Server/python-detection/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
ultralytics>=8.0.0
pillow>=10.0.0
numpy>=1.24.0
